emulators/yaze-ag: remove erroneously committed file
There were issues with the port (git had recorded some other author
after I had to reset the state to fix issues with a failed commit).
The file was the result of final consistency checks and a final
commit --amend which lead to the temp file being added to the commit.

Reported by:	danfe