math/sage: work-around to chase the upgrade of py-docutils to 0.17
There is no subdir hu for Hungarian under $PYTHON_SITELIBDIR/docutils/languages
and since the upgrade to textproc/py-docutils to 0.17 this causes a failure in
dochtml when building Sage.

Neither the upgrade to 0.17.1 proposed in PR 255197 solves the problem: as a
work-around this modification only removes the build of a_tour_of_sage for hu.