KDE Applications: sanitize DOCS option
If a port has no documentation, it doesn't need a DOCS
option (which turns on, or off, exactly an empty set
of documentation). This is now handled correctly by
kde.mk since ports 620d7678c8.

KDE **applications** with a DOCS option get doctools_build
added for free; this covers the majority of KDE-things-with
documentation. Note that DOCS isn't expected to **change**
anything: most of the ports that have (or had) that option
didn't do anything with it and always built the docs anyway.

PR:		255260
Reported by:	Naram Qashat