www/py-cherrypy: Fix runtime with Python 3.8 and onward
Backport patches from the 18.x release branch that uses the Python's
library "contextlib" instead of the external library "contextlib2".

The "contextlib2" library is a backported version which is required for
compatibility with Python 2.7 as the 17.x release of CherryPy is the
last one that has compatibility for Python 2.7.

Since the support for Python 2.7 has been removed in aa7a6c429ea6,
"contextlib2 is no longer needed for backward compatibility.

This is a transitional solution to fix the regression with Python 3.8
because an update to the 18.x release needs a bit more work and also
requires one or more new Python packages.

While I'm here:

* Sort variables according to the PHB and pet portclippy.
* Remove superfluous EXAMPLESDIR.
* Convert to newer options framework.
* Set NO_ARCH as there are no architecture specific files.
* Add "do-test" target to make future QA easier.

PR:		250379
Reported by:	jbeich
Approved by:	maintainer timeout (6+ months)
MFH:		2021Q2