security/shibboleth-sp: Update to 3.2.2
This is a security fix for an issue that has not yet been disclosed. The
vuxml entry will be updated once the CVE is available.

The patch to mitigate the vulnerability was introduced already on
2021-04-23 in the FreeBSD port as 3.2.1_1.

Security:	e4403051-a667-11eb-b9c9-6cc21735f730

(cherry-picked from commit 19889886e54f734191490f09765d631b9b67f0a7)