[Qt, GNOME] Handle os-release a little more smartly
Synth -- and therefore not all FreeBSD installations that *could*
have an /etc/os-release, or that *ought* to have an os-release
file at runtime -- doesn't have the RC script installed, so
was triggering the "depend on ports os-release" clause.

Partly revert, to checking for existence, but **also** assume
that sufficiently-new 13 (or later) releases have os-release in
base. This means that 12.2 (which ought to have it) will rely
on base, while 12.1 and 11.4 and other legacy releases will
rely on the port.

PR:	255354 251073