Uses/cabal.mk: Improve cabal-extract for USE_GITHUB/USE_GITLAB cases.
The `cabal-extract` is an auxiliary target that is intended to be run
interactively by the port developer. It populates Cabal cache and then tries
to fetch the main distribution file from Hackage package database. However,
some packages reside on GitHub and other hostings, in which case this target
fails. This commit modifies the target in such way that it does the usual
`make extract`, but preserves the Cabal cache.

Differential Revision: https://reviews.freebsd.org/D30014