shells/modernish: Fix build
It looks like modernish requires /dev/tty to be available during the
installation as it tries to run its install-time test suite. A solution
to that is wrapping ./install.sh with script(1) to make a TTY available.
Interestingly, if we replace script(1) with daemon(8), we can reproduce
the error present in the pkg fallout logs.