lang/gcc9-devel: Update to the 20210423 snapshot of GCC 9.3.1
This brings eight back ports for the tree, inter-procedural (IPO),
link-time (LTO) optimizers, three for the middle end, four for the
low level (RTL) optimizers, six related to debug information, two
for aarch64, one for x86, two related to OMP, five for the C, twentyone
for the C++ front end and 32 for libstdc++.

In other words, if you're using C++ this is a good one to consider
and test.

files/patch-libcpp_lex.c now came in via upstream, so we can (and
actually have to) remove it.