x11/plasma5-plasma-workspace: add helper script for Plasma Wayland
There are a handful of environment variables that need to be
set for a KDE Plasma Wayland session. In addition, it needs a
ConsoleKit session and its own session DBus. Add a suitable
shell script that can live alongside the regular executable.

Note that the .desktop file for Plasma does **not** use this
script -- but then again, SDDM doesn't hand off to Wayland nicely
either, so you need to start Plasma Wayland from a text console
anyway. That's simple enough now: `startplasma-wayland.sh`