deskutils/charmtimetracker: fix idle detection on Wayland
Fetch an upstream patch to avoid a crash when using charmtimetracker
on Wayland. While this avoids the crash, there's no idle-check
any more, so time tracking is less convenient than it is on X11
(where you can get up and go for lunch and deal with the idle
time later).