x11-wm/twm: generate synthetic WM hints when needed
This is based on a patch from Reddit -- the patch on Reddit is in
context-diff form and contains a bunch of cruft, so it's been
re-done (in add_window.c ; the util.c part is used unchanged).
Tested by running Chrome in previous twm -- no drop-downs at
all -- and in this lightly patched one, where it does behave.

PR:		252873
Obtained from:	Reddit