multimedia/mlt[-qt5]: update to last 6-series release
We missed a couple of 6-series releases, so jump from 6.22
straight to 6.26.1, which is the last of the 6-series.
Release notes are at:
	https://www.mltframework.org/

VDPAU support was removed -- ports had an option for it,
which was disabled and now the option is removed as well.

The 7-series introduces API breaks, so that update will
be more involved.