print/tex-luatex: Avoid mixing libc++ and libstdc++, and don't use -fcommon
This port was not respecting the CC and CXX variables, thus it could
pick up gcc and g++ instead, and that would cause mixups between
libstdc++ (which is preferred by gcc) and libc++ (which is in base),
ultimately leading to link errors.

Also, the port uses a very old version of tex so it suffers from a few
-fno-common issues, mostly variables that have been declared without an
'extern' specifier. Add a few patches to fix these.

Appproved by:	maintainer timeout
PR:		234221
MFH:		2021Q2