KDE_INVENT: drop WRKSRC hack
When USE_GITLAB was changed to follow upstream GitLab's change
in the way tag-tarballs were delivered, a hack was added to
KDE_INVENT because **that** GitLab installation lagged behind
(by one minor revision, Ithink). Now KDE's installation has
been upgraded, it had the same tag-tarball layout as other
GitLabs and the hack is no longer needed.

The ports that use KDE_INVENT have their distinfo updated,
but the ports themselves do not change so I'm not bumping
PORTREVISION; it's the same git checkout, just with different
icing on top.