devel/gnulibiberty: bring up-to-date wrt devel/binutils
Match the versions of libiberty and binutils. These come
from the same tarball upstream, so it's a bit odd to have
mismatched versions to start with; in addition the libiberty
version is years out-of-date.

While here, drop a CONFLICTS line that doesn't make sense anymore.

PR:		255794
Approved by:	joerg@ (maintainer)