sysutils/u-boot-*: Remove ubldr support
We have been using loader.efi on armv7 for a long time now.
Remove support for booting with ubldr and the needed patches that were
never upstreamed.
While here add CONFIG_EFI_GRUB_ARM32_WORKAROUND=y in the Fragment as it's
needed to have the cache flushed for us when loader.efi is started.