emulators/qemu-user-static{,-devel}: fix build on recent FreeBSD
This updates both to 6ffee39fe2e4, which pulls in some patches to fix
the build after a recent commit to main that was MFC'd to stable/13.

While we're here, update qemu-user-static-devel to point to the current
upstream. qemu-user-static-devel will likely be updated in the near
future to point to imp's 6.0 rebase branch for Q/A.

MFH:	2021Q2