New port sysutils/containerd: Open and reliable container runtime
containerd is an industry-standard container runtime with an emphasis
on simplicity, robustness and portability. It is available as a daemon
for Linux and Windows, which can manage the complete container lifecycle
of its host system: image transfer and storage, container execution and
supervision, low-level storage and network attachments, etc.

WWW: https://containerd.io/

Thanks to Samuel Karp <me@samuelkarp.com> for working on the FreeBSD
support which is all upstream already!