java/openjdk15: fix build with clang 12
During an exp-run for llvm 12 (see bug 255570), it turned out that at
least openjdk11 and openjdk12 do not build with clang 12.0.0. The
exp-run therefore skipped openjdk15.

Building this manually shows that it results in a compile error:

gmake[4]: Leaving directory '/wrkdirs/usr/ports/java/openjdk15/work/jdk15u-jdk-15.0.2-7-1/make'
/wrkdirs/usr/ports/java/openjdk15/work/jdk15u-jdk-15.0.2-7-1/src/hotspot/cpu/x86/vm_version_ext_x86.cpp:748:3: error: suspicious concatenation of string literals in an array initialization; did you mean to separate the elements with a comma? [-Werror,-Wstring-concatenation]
  "",
  ^
/wrkdirs/usr/ports/java/openjdk15/work/jdk15u-jdk-15.0.2-7-1/src/hotspot/cpu/x86/vm_version_ext_x86.cpp:747:3: note: place parentheses around the string literal to silence warning
  "Opteron QC/Phenom"  // Barcelona et.al.
  ^
1 error generated.

This is due to a missing backport of this upstream commit:

commit f8a9602a0a65cdc98eb940aac9529256ded2bf42
Author: Yasumasa Suenaga <ysuenaga@openjdk.org>
Date:	Thu Jan 21 06:08:13 2021 +0000

    8260025: Missing comma in VM_Version_Ext::_family_id_amd

    Reviewed-by: dholmes, stuefe

Even after applying this fix, there is still a possibility of a segfault
during the build, due to another missing backport, of this upstream
commit:

commit c484d8904285652246c3af212a4211b9a8955149
Author: Thomas Stuefe <stuefe@openjdk.org>
Date:	Tue Mar 16 05:49:01 2021 +0000

    8263557: Possible NULL dereference in Arena::destruct_contents()

    Reviewed-by: kbarrett, coleenp

Approved by:	maintainer timeout (2 weeks)
PR:		255904
MFH:		2021Q2