www/pydio-cells: Fix build and overhaul
Summary:
Pydio cells uses a custom version of packr to embed the static
resources into the `cells` binary. Previously, the embedding process
was not peformed, resulting in a non-functional binary. When trying
to use the app, user will just get "404 Not Found" error instead of
a web page.

To fix this - before the build `gmake generate` has to be performed
in the WRKSRC, and a special custom version of packr v1
(devel/pydio-packr) has to be used.

Other changes:
- DOCS option was removed, as the README.md file didn't really provide
  any useful information for the port's users
- Generation of bash and zsh completion files was added
- App version is now provided during the build process (previously 0.1.0
  would be indicated instead of the real version)
- Indicate failure to build on i386
- Create separate user/group for running the daemon, as recommended by
  the administrator's manual
- Add rc script
- Add pkg-message with configuration instructions
- Add more patches fixing bugs and adapting the code for FreeBSD

No UPDATING is required, as the port was not functional previously

Reviewed by:	koobs
MFH:		No (port isn't in any of the quarterly branches yet)
Differential Revision:		https://reviews.freebsd.org/D30323