devel/rubygem-drb: Add rubygem-drb 2.0.4
dRuby is a distributed object system for Ruby. It allows an object in one Ruby
process to invoke methods on an object in another Ruby process on the same or a
different machine.

The Ruby standard library contains the core classes of the dRuby package.
However, the full package also includes access control lists and the Rinda
tuple-space distributed task management system, as well as a large number of
samples.

WWW: https://github.com/ruby/drb