Mk/bsd.port.mk: Don't pass up an empty flavor.
The classic way to handle flavors is to set:

  FLAVORS= foo bar
  FLAVOR?= ${FLAVORS:[0]}

And in that case, FLAVOR is only set if the variable is not defined. If
you pass an empty flavor using `FLAVOR=` then it remains empty after
that line.  It can leads to some ports with complicated logic to assume
the wrong flavor is set.

PR:		256301
Reported by:	avg
Differential Revision:	https://reviews.freebsd.org/D30579