x11/konsole: allow /bin/sh as shell
This pulls in an upstream fix (developed by Kurt Hindenburg
and myself) to correct a logic error that disallows /bin/sh
as shell in konsole sessions.

While here, reduce dependencies a little (helloSystem might
like that) because ecm is a build-dep and only very rarely
a runtime-dep; kdelibs4support isn't needed at all.