print/fig2dev: Add license text
The text of this license has been extracted from: configure.ac
and refers to all of fig2dev. It includes a list of historic authors
and contributors, and while not resembling a standard license, it is
clear that all rights to build, use, and distribute this software
without any restrictions is granted.