graphics/xfig: update to upstream version 3.2.8a
The port update to version 3.2.8 has been provided in PR 252092 by
mjl@luckie.org.nz. This PR will not be closed by this commit, since
it did also contain a port update for the print/fig2dev port.

I have adjusted the update to build patch level "a" of version 3.2.8
and have added missing dependencies and applied a few style changes
to the Makefile as suggested by portlint and portclippy.