category/port: Package name dependencies do not support globbing.
5.9.9 of the Porters Handbook defines the use of pkgname/version
dependencies which does not include globbing.  Incidentally
`pkg info` is used by the ports framework which does support it
but external tools, like Poudriere, don't expect this.

None of these needed globbing.  If there is a good reason beyond
convenience for the feature then it needs to be documented and
tools updated.