games/bubble-chains: resurrect port and update to 0.2.0
Bubble Chains is a funny 2D game which runs on Linux, Windows and
FreeBSD.

You have to collect color bubbles on the field by combining them
into chains of three or more bubbles. Collected chains disapper,
making other bubbles falling down. The longer is a chain, the more
score is added to the corresponding target at the bottom. Completed
targets are then removed from the field.

The aim of Bubble Chains is to remove all of the targets on each
level, and to do this before the time (indicated with the blue-colored
bar at the right) runs out. Keep doing this until you have passed
the last level and won the game.

WWW: https://github.com/ArsMasiuk/bubble-chains