math/R-cran-wk: Add new port
Provides a minimal R and C++ API for parsing well-known binary and
well-known text representation of geometries to and from R-native
formats. Well-known binary is compact and fast to parse; well-known
text is human-readable and is useful for writing tests. These formats
are only useful in R if the information they contain can be accessed
in R, for which high-performance functions are provided here.

WWW: https://cran.r-project.org/web/packages/wk/