Mk/Uses: replace bare ${ECHO} with ${ECHO_CMD} or ${ECHO_MSG}
ECHO is builtin variable and is cleared when make(1) is invoked
in silent mode, i.e. as ``make -s'', thus making statements that
use it do nothing.  Depending on the context, replace it with
either ${ECHO_CMD} or ${ECHO_MSG} (in one case), as appropriate.

PR:		256185
Submitted by:	Franco Fichtner