devel/libsoup: make samba a (not registered) runtime dep
This picks up an idea from PR 235207 again, where NTLM authentication
is optional. Previous commits made samba a build-dependency so
that the --enable-ntlm option would work. The meson build -- even
though the dep is entirely runtime -- would check for the presence
of ntlm_auth at build-time, to be able to pass in the path to
ntlm_auth to the build.

Re-jig this to not do the check, accept the option as-is, and
use a likely path to ntlm_auth instead (which is what the original
patch in 235207 did).

A circular dependency from samba41[23] -> libsoup and back
can be triggered by setting option SMB in libsoup and
option SPOTLIGHT in samba. Originally the runtime dependency
was registered as
	USES=	samba:run
A _RUN dependency still wants to build, though, and package and
index builds failed. Cut the circular dependency by **not**
registering samba as a runtime dependency and using a pkg-message
to tell users to possibly install samba.

Anyone who now has "but my samba is installed in a different prefix"
can file a PR with a patch fixing meson better.

PR:		256076
Reported by:	Dimitry Andric