mail/postfix: Revert to 3.5.9
Revert recent update to 3.6.1 back to 3.5.9 to fix the build.  As there
is no mail/postfix35 that builds on FreeBSD 11, keep mail/postfix on the
last version known to build and run on FreeBSD 11.

And while here, also update the MAINTAINER. No functional change.

This is a direct commit to 2021Q2.