devel/py-qt5-positioning: New port - Provide PyQt5 QtPositioning module
The QtPositioning module contains classes to determine a position by
using a variety of possible sources, including satellite, or wifi, or
a text file, and so on. That information can then be used to, for
example, determine a position on a map. In addition satellite
information can be retrieved and area based monitoring can be performed.

https://www.riverbankcomputing.com/static/Docs/PyQt5/api/qtpositioning/qtpositioning-module.html

Reviewed by:	lbartoletti, #portmgr!

Approved by:	lbartoletti (kde@)

Differential Revision:	https://reviews.freebsd.org/D30829