x11/nwg-menu: add new port
nwg-menu displays the system menu with simplified FreeDesktop Main
Categories (8 instead of 13). It also provides the search entry, to
look for installed application on the basis of .desktop files, and for
files in XDG user directories.

It's being developed with sway in mind, but should also work with
other wlroots-based Wayland compositors such as Wayfire.

https://github.com/nwg-piotr/nwg-menu