biology/sra-tools: NCBI's toolkit for handling data in INSDC Sequence Read Archives
SRA tools is a toolkit for using data in the INSDC Sequence Read Archives.

SRAs operated by International Nucleotide Sequence Database Collaboration
houses sequence reads and alignments generated by "next-gen" sequencers.

This port is a bit convoluted due to the fact that the sra-tools build
requires access to the ncbi-vdb source tree.  Hence, ncbi-vdb is treated
as a submodule here rather than a separate library port.  We are working
with upstream with hope for long-term improvements.