math/maxima: Fix staging error
configure.ac uses the git command if present to populate the VERSION variable.
As a consecuence, this produces a dynamically changing installation directory in
STAGEDIR that breaks the build in some conditions. Note that on a clean
environment (poudriere), this works fine.

This patch comments that portion of configure.ac out since it is of no use
inside the ports tree.

PR:	255683
Reported by:	russo@bogodyn.org