biology/linux-foldingathome: Update to 7.6.21
Unfortunately, this version of FAHClient is not able to daemonize
successfully anymore. Even though the --daemon flag is passed, FAHClient
does not detach from the controlling terminal. In addition to that it
does not go into background. The --log seems to be broken as well as
since all the logs are printed to the terminal instead of the log file.

In order to alleviate those problems, use daemon(8).

PR:		250463
Changes:	https://foldingforum.org/viewtopic.php?t=36307