security/nebula: New port
Nebula is a scalable overlay networking tool with
a focus on performance, simplicity and security.
It lets you seamlessly connect computers anywhere
in the world. Nebula is portable, and runs on Linux,
OSX, Windows, iOS, and Android. It can be used to
connect a small number of computers, but is also
able to connect tens of thousands of computers.

WWW: https://github.com/slackhq/nebula