x11/nwg-bar: add new port
nwg-bar is a Golang replacement to the nwgbar command (a part of
nwg-launchers), with some improvements. Aimed at sway, works with
wlroots-based compositors only.

The nwg-bar command creates a button bar on the basis of a JSON
template placed in the ~/.config/nwg-bar/ folder.  By default the
command displays a horizontal bar in the center of the screen. Use
command line arguments to change the placement.

https://github.com/nwg-piotr/nwg-bar