x11-servers/xorg-server: patch to utilize CLOCK_MONOTONIC_FAST for the clock
This patches the X server to use CLOCK_MONOTONIC_FAST instead of
CLOCK_MONOTONIC for the clock.  This is basically a reintroduction of
r208485 by jkim@ that was removed by accident(?).

Since the X server is fine with the Linux _COARSE clock the accuracy of
the FreeBSD _FAST clock should work too.  I have run with this patch for
a long time without ill effects.

PR:		252816
Reviewed by:	imp, markj, jkim, mhorne
Approved by:	x11