graphics/poppler: make boost a build-time dependency
Poppler uses Boost small-vectors for efficiency (over std::vector)
when the Splash renderer is enabled. Small-vectors are header-only,
so there is no need to depend on the whole boost-libs at runtime.

The PR asks to drop Boost entirely; we're not going to do that,
but as a build-dep it avoids installed-size bloat.

Bump PORTREVISION also of related ports to make sure they are
rebuilt; don't need to re-do all consumers because Boost was not
linked anyway.

PR:		256821