devel/p5-XS-Parse-Sublike: Add p5-XS-Parse-Sublike 0.12
XS::Parse::Sublike provides some XS functions to assist in writing parsers for
sub-like syntax, primarily for authors of keyword plugins using the
PL_keyword_plugin hook mechanism. It is unlikely to be of much use to anyone
else; and highly unlikely to be any use when writing perl code using these.
Unless you are writing a keyword plugin using XS, this module is not for you.

WWW: https://metacpan.org/dist/XS-Parse-Sublike