security/shibboleth-sp: update to 3.2.3
A regression in the RequestMap feature causing random crashes on some
systems was identified, necessitating this patch update. Most uses of
this feature tend to be on Windows, so that's the primary platform
affected but the bug was generic if the feature were to be used on other
systems.

It isn't easily exploitable by specific requests, so it's a borderline
sort of denial of service risk and the Shibboleth project chose not to
do an advisory, and anybody affected won't need much incentive to get
the patch anyway.