ports-mgmt/portmaster: make --no-confirm omit the PAGER
As requested in Github issue #64 the pager should not be used if
portmaster is executed within a script. In order to not introduce
a new option, --no-conform will now imply that output will not be
halted waiting for user input to proceed when displaying the package
messages.

Reported by:	Michael Osipov (michael-o on Github)