databases/qt5-sqldrivers-mysql: fix against mysql 5.7.34
In mysql 5.7.34 binding a parameter to a 0-sized buffer triggers
an assert() -- previously this was DBG_ASSERT. This triggers
crashes in databases/akonadi in particular, which binds blobs of
size 0. This patch allows akonadi with the default mysql to
start (at all) again.

The patch file previously held a Qt4-era patch which had long
ago had a better patch applied upstream (the comment in the
patch refers to that). The old stuff has been removed, leaving
only the parameter-fix.