security/zeek: Unbreak build under 14.0-CURRENT
According to the cpuset(2) man page, sys/param.h must be included
before sys/cpuset.h. This was fixed in zeek (in the highwayhash
submodule) in May of 2020 and undone in August of 2020. Add a patch
that matches the pull request I just created with upstream:

    https://github.com/zeek/highwayhash/pull/1

Thanks to @pluknet for diagnosing the build failure.

Reported by:	pkg-fallout