Mk/Scripts/qa.sh: Turn off pipefail when piping to grep -q
The pipeline

   readelf -d "${dep_file}" | grep -q SONAME

can fail because grep -q closes the output early resulting sigpipe being
sent to readelf.  Other possible solutions are to turn off pipefail for
the file or remove the -q grep argument.

Credit to ashish@ for the detective work to discover the root cause of
this.

Differential Revision: https://reviews.freebsd.org/D31211

Reviewed by:	emaste
Approved by:	bapt (portmgr) mat (portmgr)