security/rkhunter: use sha256 instead of sha256sum
Coreutils compatible versions of the hash commands have been added
to 14-CURRENT and 13-STABLE. These were preferred by rkhunter, when
found, but without adjustment of the position of the hash in the
generated output.

Instead of trying to adjust for different positions of the hash,
this patch removes the preference for sha256sum over sha256 and
restores the behavior this script had before sha256sum has been
made available in recent FreeBSD versions.

Tested by:	Łukasz Wąsikowski (maintainer)