Mk/bsd.port.mk: prepare the land for pkg 1.17
in next version of pkg, we now use a uniq extension: '.pkg' whatever the
compression is, the problem is some bootstrap pkg(7) does not handle
that yet, notably on freebsd 11.4 and 12.2.

We already have a compatibility layer by creating a symlink on the old
extension, which works but the signature itself will not exist with the
old extension. Add a workaround in the ports tree to symlink the new
signature to the old expected signature.

Given at the moment we are doing that things are not yet signed by
default we do create a dead symlink, but at the moment the repo are
published the right signature will be there.